refactor: fga to internal package
On-behalf-of: @SAP <simon@simontesar.com>
